fix(): redistributes constants
